starts-with? and ends-with? and contains? return true for empty sub-string.
This behavior is consistent with existing similar methods in both
Java and goog.string.
The logic is that zero length string is "contained" between any two chars.

nil in Clojure is a replacement for empty collection in many situations
so nil should be treated as empty string also.

Case when containing string is null produces NullPointerException so
returning false for (starts-with? nil nil) is probably OK but I think
it should be true.